+++ src/xulbux/cli/tools.py
@@ -3,14 +3,17 @@ from ..console import Console
 
 
 def render_format_codes():
+    """CLI command function for `xulbux-lib fc` command, which allows you to parse
+    and render a given string's format codes as ANSI console output."""
     args = Console.get_args({"input": "before"})
+    vals = args.input.values[1:]  # EXCLUDE THE COMMAND ITSELF
 
-    if not args.input.values:
+    if not vals:
         FormatCodes.print("\n[_|i|dim]Provide a string to parse and render\n"
                           "its format codes as ANSI console output.[_]\n")
 
     else:
-        ansi = FormatCodes.to_ansi("".join(args.input.values))
+        ansi = FormatCodes.to_ansi("".join(vals))
         ansi_escaped = FormatCodes.escape_ansi(ansi)
         ansi_stripped = FormatCodes.remove_ansi(ansi)